Remove support for C# dynamic from JsonNode
