+++ src/libraries/System.Text.Json/tests/System.Text.Json.Tests/System.Text.Json.Tests.csproj
@@ -32,7 +32,6 @@
     <Compile Include="JsonEncodedTextTests.cs" />
     <Compile Include="JsonGuidTestData.cs" />
     <Compile Include="JsonNode\Common.cs" />
-    <Compile Include="JsonNode\DynamicTests.cs" />
     <Compile Include="JsonNode\JsonArrayTests.cs" />
     <Compile Include="JsonNode\JsonNodeTests.cs" />
     <Compile Include="JsonNode\JsonNodeOperatorTests.cs" />
